add a progress bar (not working yet) and reorganize file structure
