+++ components/Quote/Quote.module.css
@@ -27,34 +27,3 @@
 .quoteBlock figcaption {
   color: darkgray;
 }
-
-.quoteControls {
-  background: white;
-  border: 1px solid black;
-}
-@supports (display: grid) {
-  .quoteControls {
-    display: grid;
-    grid-template-columns: 1fr;
-  }
-
-  @media screen and (min-width: 768px) {
-    .quoteControls {
-      grid-template-columns: 1fr 1fr;
-    }
-  }
-}
-.quoteControls button {
-  border: 1px solid black;
-  padding: 8px 16px;
-  background: none;
-  font-weight: 600;
-}
-.quoteControls button[aria-checked="true"] {
-  background: blue;
-  color: white;
-}
-
-.progressBar {
-  background: blue;
-}
