+++ components/QuoteControls/QuoteControls.module.css
@@ -1,0 +1,26 @@
+.quoteControls {
+  background: white;
+  border: 1px solid black;
+}
+@supports (display: grid) {
+  .quoteControls {
+    display: grid;
+    grid-template-columns: 1fr;
+  }
+
+  @media screen and (min-width: 768px) {
+    .quoteControls {
+      grid-template-columns: 1fr 1fr;
+    }
+  }
+}
+.quoteControls button {
+  border: 1px solid black;
+  padding: 8px 16px;
+  background: none;
+  font-weight: 600;
+}
+.quoteControls button[aria-checked="true"] {
+  background: blue;
+  color: white;
+}
