+++ utils/percentage.ts
@@ -1,0 +1,3 @@
+export function percentage(value: number, max: number): number {
+  return (value / 100) * max;
+}
